Put LiveStreamUpdater under the Coronavirus namespace
This is for consistency with the rest of the live stream code.
